fixed message id and message server

+++ src/main/java/core/model/Request.java
@@ -6,5 +6,5 @@ import com.fasterxml.jackson.annotation.JsonProperty;
  * Messaging request record.
  */
 public record Request(Command command, @JsonProperty("peer_id") String peerId,
-        @JsonProperty("message_id") Integer messageId, String message) {
+        @JsonProperty("message_id") Long messageId, String message) {
 }
